requirements.txt
daphne==2.4.1
Django==3.0.4
django-debug-toolbar==2.2
django-debug-toolbar-template-profiler==2.0.0
django-extensions==2.2.8
django-allauth==0.41.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
Pillow==7.0.0
pipupgrade==1.6.8
python-telegram-bot==12.4.2